Centralize config and split Neural_Network.py into nn/ package
Pulls hardcoded paths, hyperparameters, YOLO confidence thresholds, FPS,
and the Postgres URL into a single config.py so stage scripts no longer
redeclare the same env-var boilerplate or magic numbers in-place.

Splits Neural_Network.py (1601 lines) into a small nn/ package -
dataset, transforms, model, training, pruning, plots, io plus a shared
_logger - and leaves Neural_Network.py as a thin entrypoint that wires
config to the package. Behavior is preserved: same SEED, same logger
name, same fallback-path semantics on dir-creation failure, same
augmentation pipeline.

Object_Tracking / Feature_Engineering / Data_Loading now read their
tunables (YOLO thresholds, FPS, DB URL) from config rather than from
literals in function defaults.
